requirements.txt
Django==2.2.7
djangorestframework==3.11.0
Pillow==6.2.1
psycopg2==2.8.4
python-dotenv==0.10.3
pytz==2019.3
sqlparse==0.3.0
